+++ reactos/lib/msvcrt/mbstring/hanzen.c
@@ -8,8 +8,8 @@
 		Modified from Taiji Yamada japanese code system utilities
  *              12/04/99: Created
  */
+#include <msvcrti.h>
 
-#include <msvcrt/mbctype.h>
 
 static unsigned short han_to_zen_ascii_table[0x5f] = {
   0x8140, 0x8149, 0x8168, 0x8194, 0x8190, 0x8193, 0x8195, 0x8166,
@@ -61,7 +61,7 @@ static unsigned char zen_to_han_symbol_table_2[ZTOH_SYMBOLS] = {
 #define JTOKANA(c) ((c) <= 0x82dd ? (c) + 0xa1 : (c) + 0xa2)
 
  
-unsigned short _mbbtombc(unsigned short c)
+unsigned int _mbbtombc(unsigned int c)
 {
   if (c >= 0x20 && c <= 0x7e) {
     return han_to_zen_ascii_table[c - 0x20];
@@ -72,7 +72,7 @@ unsigned short _mbbtombc(unsigned short c)
 }
 
 
-unsigned short _mbctombb(unsigned short c)
+unsigned int _mbctombb(unsigned int c)
 {
   int i;
   unsigned short *p;
