+++ reactos/lib/msvcrt/mbstring/jistojms.c
@@ -1,6 +1,7 @@
-#include <msvcrt/mbstring.h>
+#include <msvcrti.h>
 
-unsigned short _mbcjistojms(unsigned short c)
+
+unsigned int _mbcjistojms(unsigned int c)
 {
   int c1, c2;
